WebClient samples renamed
