+++ examples/webclient/standalone/pom.xml
@@ -25,8 +25,8 @@
         <version>2.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>helidon-examples-webclient-basic-example</artifactId>
-    <name>Helidon WebClient Basic Example</name>
+    <artifactId>helidon-examples-webclient-standalone</artifactId>
+    <name>Helidon WebClient Standalone Example</name>
 
     <dependencies>
         <dependency>
