+++ examples/webclient/standalone/src/main/java/io/helidon/examples/webclient/standalone/package-info.java
@@ -16,4 +16,4 @@
 /**
  * Basic examples of webclient usage.
  */
-package io.helidon.webclient.example;
+package io.helidon.examples.webclient.standalone;
