+++ examples/webclient/standalone/src/main/resources/application.yaml
