+++ examples/webclient/standalone/src/main/resources/full-webclient-config.yaml
